FEAT: planet generation start

+++ backend/src/app/database/database_access/planet_access.py
@@ -39,7 +39,7 @@ class PlanetAccess:
         planet_id = planet.id
         return planet_id
 
-    async def createPlanetRegion(self, planet_id: int, region_type: str, coordinate: tuple[int, int]):
+    async def createPlanetRegion(self, planet_id: int, region_type: str, coordinate: tuple[float, float]):
         """
         create a region on a planet
